Split collapse plugin out of alert.jsx file
