+++ src/alert.jsx
@@ -1,36 +1,5 @@
-import React, { Component } from "react";
-
-class Collapse extends Component {
-	constructor(props) {
-		super(props);
-
-		this.state = {
-			open: !!this.props.open
-		};
-	}
-
-	componentDidMount() {
-		this.setState({
-			open: true
-		});
-	}
-
-	componentWillUnmount() {
-		this.setState({
-			open: false
-		});
-	}
-
-	render() {
-		return (
-			<div {...this.props} style={{
-				overflow: "hidden",
-				maxHeight: this.state.open ? "999em" : "0",
-				transition: ".25s height"
-			}}>{this.props.children}</div>
-		);
-	}
-}
+import React from "react";
+import Collapse from "./collapse";
 
 const Alert = ({ type = "info", message, headline, onDismiss }) => {
 	const css = "alert alert-dismissible alert-" + type;
